+++ xcube/core/byoa/constants.py
